PW_FT_classification/requirements.txt
PytorchWildlife
scikit_learn
lightning
munch
typer